factor wire protocol into module and add tests
